Added initial support for SysFS GPIO class.
Nify thing that is NON-RAII (Why would you act like you just leaked
out memory on a crash with GPIOs...  Amazing BURNING STUPIDS there-
and it's in one the GPIO C++ classes for Linux...  X-D)

It lets you encapsulate operations on the GPIO as specified- which
is to say, you can set IN/OUT, read IN values, reset the IN value
so it goes live, read/set OUT values, and set callbacks against
GPIO line events where the state changes.  Does this CLEANLY
unlike most of the others, with as minimal an overhead as
is possible in C++ without using convoluted C code wrapped into
classes.

Default is OFF for now.  Will revisit when we can extend this
more to supporting Windows and OSX.