Added support/distinction of controlling sense of GPIO...
On open/init, we used to force it active high and force the line down-
for LEDs that're set active low, this makes them blink until the code
clamps it the "right" direction.  So, we're going to let it be
invertable, with the option for that to be active high and you
can override it accordingly on construction.