Re-worked Runnable a bit...
Runnable/OneShot needed to be re-worked a bit.  Tools has become a C++11
type classlib really- with it's features being adjustable to what the
compiler can manage- if Boost is available and not a footprint problem,
use it if the standard stuff doesn't work right.  If threading doesn't
work right (not as much of an issue these days) tinythread++ is
readily available when Boost isn't for those cases for threading, etc.

So...claiming no C++11 is a bit bogus.  So we're going to slowly
clean this up and make it all really clear, leveraging C++11 stuff
so that we can leverage things like NONCOPY as a base class.

So, I removed a bit of cruft from the old way of doing things- which
is deprecated if you're using C++11 or later.  I also caught an
initalization fubar that my past use-cases didn't trip across.