Further cleanups on Runnable.
Since this was using tinythread in a prior incantaion (and I could
still go back there, if I had to...), there's a few sins in here
that're clearly ones for C++11 threading rules that Runnable needs
anyhow.  Use cases that I am now needing that weren't tested for
but are still valid for what this class is supposed to do.