Made the GPIO controller class a bit smarter about teardown.
If it existed before we came along to init, we shouldn't be doing
teardowns.  Care should be exercised for GPIOs so that you're not
yanking a carpet out from underneath another object using the same
GPIOs, etc. but this should be the simplest means to ensure you're
not wiping out the GPIO hooks for another instance at the same GPIO
or an app using this or another engine for the same.